s,u, a normer

+++ ft_convert_percentage_d.c
@@ -6,7 +6,7 @@
 /*   By: mzhan <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/29 12:03:57 by mzhan             #+#    #+#             */
-/*   Updated: 2021/04/08 17:41:59 by mzhan            ###   ########.fr       */
+/*   Updated: 2021/04/09 14:01:04 by mzhan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,7 +87,7 @@ void	ft_prec_sup_len_global(t_struct *f, int res, int len)
 
 void	ft_convert_percentage_d(va_list *arguments, t_struct *f)
 {
-	long int	res;
+	int			res;
 	int			len;
 
 	res = va_arg(*arguments, int);
